Refactor YDTB Widget
